+++ CHANGELOG.md
@@ -1,6 +1,9 @@
-Unreleased
-----------
+0.18.0
+------
+- Introduced trailing stop order types
 - Added support for paging of account activity data
+- Added support for specifying reported account activity direction as
+  well as `until` and `after` times
 - Bumped `time-util` dependency to `0.3`